requirements.txt
opencv-python
tensorflow
keras
requests
urllib3
python-dotenv
vk_api
